🚚 Moved OG function into library
